所有double 替换成float

+++ src/data.hpp
@@ -53,7 +53,7 @@ public:
     std::vector<std::vector<svgPoint *>> data;      //原始的数据
     std::vector<std::vector<std::vector<unsigned>>> propagator;
     std::vector<AbstractFeature> feature;                             //图案数据
-    std::vector<double> features_frequency;                            //图案频率
+    std::vector<float> features_frequency;                            //图案频率
 
 };
